website/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==4.1
django-jazzmin==3.0.1
django-sendgrid-v5==1.2.4
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
sendgrid==6.11.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
